Move transforms to sample apps
