+++ sample-apps/radiology/lib/transforms/transforms.py
@@ -1,0 +1,46 @@
+# Copyright (c) MONAI Consortium
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#     http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import logging
+from typing import Dict, Hashable, Mapping
+
+import numpy as np
+from monai.config import KeysCollection
+
+# from monai.networks.layers import GaussianFilter
+from monai.transforms.transform import MapTransform
+
+logger = logging.getLogger(__name__)
+
+
+class HeatMapROId(MapTransform):
+    def __init__(
+        self,
+        keys: KeysCollection,
+        allow_missing_keys: bool = False,
+    ):
+        """
+        Convert to single label - This should actually create the heat map for the first stage
+
+        :param keys: The ``keys`` parameter will be used to get and set the actual data item to transform
+
+        """
+        super().__init__(keys, allow_missing_keys)
+
+    def __call__(self, data: Mapping[Hashable, np.ndarray]) -> Dict[Hashable, np.ndarray]:
+        d: Dict = dict(data)
+        for key in self.key_iterator(d):
+            if key == "label":
+                # Convert to single label
+                d[key][d[key] > 0] = 1
+            else:
+                print("This transform only applies to the label")
+        return d
