+++ sample-apps/radiology/main.py
@@ -265,7 +265,7 @@ def main():
     app.train(
         request={
             "model": args.model,
-            "max_epochs": 500,
+            "max_epochs": 2000,
             "dataset": "CacheDataset",  # PersistentDataset, CacheDataset
             "train_batch_size": 1,
             "val_batch_size": 1,
